+++ release-notes.rst
@@ -5,14 +5,14 @@ Release notes for chevah.utils
 0.20.1 - 21/07/2013
 -------------------
 
-* Fix setting special boolean value.
+* Fix setting special (inherit) boolean value.
 
 
 0.20.0 - 21/07/2013
 -------------------
 
-* Don't set bad value for configuration options and raise an error when
-  setting bad value.
+* Don't set invalid values for configuration options and raise an error when
+  trying to set one.
 
 
 0.19.2 - 18/07/2013